Separated styled letter file, renamed vscode auto named TSXs to TSes
