+++ src/components/__styled__/StyledWord.ts
